\name{BAMMlikelihood}
\alias{BAMMlikelihood}
\title{
Calculate BAMM likelihood
}
\description{
Calculates the likelihood of a phylogeny exactly as is done by BAMM, given a set of events.
}
\usage{
BAMMlikelihood(phy, eventdata, gen = 'last', segLength = 0.02, sf = 1)
}
\arguments{
  \item{phy}{
either an object of class \code{phylo} or the path to a tree file in newick format.
}
  \item{eventdata}{
a table of event data, as returned by BAMM, either as an object of class \code{dataframe} or as the path to an event_data file. Alternatively, a named numeric vector of length two holding speciation ("lambda") and extinction ("mu") rates for the constant-rate birth-death process.
}
  \item{gen}{
the BAMM generation for which the likelihood should be calculated. Can be an integer specifying a specific generation, or \code{last}, specifying the last generation, or \code{all}, in which case the likelihood will be calculated for all generations. 	
}
  \item{sf}{
sampling fraction.   	
}
\item{segLength}{
The relative segment length, exactly as defined for BAMM   	
}
\item{e_prob_condition}{
How to handle extinction probabilities at internal nodes when right and left descendant branches have different shift histories. The default in BAMM is 'if_different' and we do not recommend changing this.
}

}
\details{
This function allows the user to check the likelihoods computed by BAMM using an independent R-based implementation. This is designed to provide a check on potential software bugs that might be introduced during future BAMM development and which might compromise the likelihood calculation. If you observe measurable discrepancies between the likelihood computed by this function and the corresponding likelihood returned by BAMM, please inform the BAMM development team.
}

\author{
Dan Rabosky, Pascal Title
}

\examples{
library(BAMMtools)
data(whales, events.whales, mcmc.whales)

x <- BAMMlikelihood(whales, events.whales, gen = 'last')

# Does the likelihood generated by BAMM match the R implementation?
identical(round(x, 3), mcmc.whales[nrow(mcmc.whales), 'logLik'])

# an example with a constant-rate birth-death process:
pars <- c(0.5, 0.45)
names(pars) <- c("lambda", "mu")
BAMMlikelihood(whales, pars)
 
}
