\name{BAMMtools-internal}
\alias{arc}
\alias{areEventConfigurationsIdentical}
\alias{areShiftSetsEqual}
\alias{as.bammdata}
\alias{as.bammdata.credibleshiftset}
\alias{as.phylo.bammdata}
\alias{as.phylo.branchprior}
\alias{barLegend}
\alias{branchMeanRateExponential}
\alias{buildSegmentMatrix}
\alias{buildSegmentMatricesRecursive}
\alias{buildTreeWithEventList}
\alias{colorMap}
\alias{computeBAMMlikelihood}
\alias{D_func}
\alias{E_func}
\alias{eventMatrix}
\alias{exponentialRate}
\alias{getPathToRoot}
\alias{getRecursiveSequence}
\alias{getSampleCoMat}
\alias{getSequenceForwardTraversal}
\alias{getStartStopTimes}
\alias{rateLegend}
\alias{inv.logit}
\alias{logit}
\alias{makeBDeventMatrix}
\alias{meanExponentialRate}
\alias{mkdtsegsPolar}
\alias{mkdtsegsPhylo}
\alias{NU.branching.times}
\alias{richColors}
\alias{recursiveAddEventsToTree}
\alias{recursivelySetNodeStates}
\alias{safeLog}
\alias{segMap}
\alias{timeIntegratedBranchRate}
\alias{transparentColor}
\alias{getDesc}
\alias{getSpanningTaxonPair}
\alias{print.bammdata}
\alias{setPhyloTreeCoords}
\alias{setPolarTreeCoords}
\alias{phylogeneticMean}
\alias{palettes}
\alias{redirect}

\title{Internal BAMMtools functions}
\description{
\code{BAMMtools-internal} functions are not routinely called by the user and are undocumented.
}

\details{
Internal \pkg{BAMMtools} functions, either not intended to be called directly by the user or otherwise undocumented.
}
 