\name{cohorts}
\alias{cohorts}
\title{
Visualize macroevolutionary cohorts
}
\description{
Plots the matrix of pairwise correlations in rate regimes between all tips in a phylogeny
}
\usage{
cohorts(x, ephy, col, pal, lwd = 1, use.plot.bammdata = TRUE, useraster = FALSE)
}
\arguments{
  \item{x}{
a matrix of pairwise correlations generated by \code{getCohortMatrix}.
}
  \item{ephy}{
an object of class \code{bammdata}.
}
  \item{col}{
a vector of colors passed to the function \code{image}. These will be used
to color the values in \code{x}. See documentation for \code{image}.  	
}
  \item{pal}{
the palette to use if \code{use.plot.bammdata=TRUE}. See options documented in
the help file for \code{plot.bammdata}.  	
}
  \item{lwd}{
a numeric indicating the width of branches in the phylogeny.  	
}
  \item{useraster}{
a logical indicating whether the function \code{image} should plot the matrix 
as a raster.  	
}
}

\author{
Mike Grundler
}

\seealso{
\code{\link{plot.bammdata}}, \code{\link{getCohortMatrix}}, \code{\link{image}}
}
\examples{
library(BAMMtools)
data(whales, events.whales)
ed <- getEventData(whales, events.whales, 0.1)
x <- getCohortMatrix(ed)
cohorts(x, ed)
cohorts(x, ed, col=richColors(64))
cohorts(x, ed, pal="temperature", col=richColors(64))
}
