% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bammcolors.R
\name{richColors}
\alias{richColors}
\title{Rich color palette}
\usage{
richColors(n)
}
\arguments{
\item{n}{The number of desired colors.}
}
\description{
Deprecated function. Please use
    \code{\link[gplots]{rich.colors}} instead.
}
\seealso{
\code{\link[gplots]{rich.colors}}
}
