\name{summary.branchprior}
\alias{print.branchprior}
\title{
Summary of class "branchprior" object.
}
\description{
Prints summary attributes of the \code{branchprior} object 
}
\usage{
summary(ox)
print(ox)
}
\arguments{
  \item{ox}{
an object of class \code{branchprior}
}

}
\details{
 
}
\value{
None
}

\references{
\code{bamm-project.org}
}
\author{
Dan Rabosky
}

\seealso{
\code{\link{distinctShiftConfigurations}}, \code{\link{plot.bammshifts}}, \code{\link{credibleShiftSet}}\code{\link{getBranchShiftPriors}}
} 
 
 
\examples{
 
}
 
\keyword{ models }
