% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparentColor.R
\name{transparentColor}
\alias{transparentColor}
\title{Define colors with transparency}
\usage{
transparentColor(namedColor, alpha = 0.8)
}
\arguments{
\item{namedColor}{A color name.}

\item{alpha}{A transparency value between 0 and 1, where 0 is fully
transparent.}
}
\value{
Returns the transparent color in RGB format.
}
\description{
Converts a named color and opacity and returns the proper RGB
    code.
}
\details{
This function is used internally by
    \code{\link{plotRateThroughTime}}.
}
\author{
Pascal Title
}
\keyword{manip}
