\name{BAMMtools-internal}
\alias{segMap}
\alias{decimals}
\alias{inv.logit}
\alias{logit}


\title{Internal BAMMtools functions}
\description{
\code{BAMMtools-internal} functions are not routinely called by the user and are undocumented.
}

\details{
Internal \pkg{BAMMtools} functions, either not intended to be called directly by the user or otherwise undocumented.
}
 