\name{BAMMtools-internal}
\alias{arc}
\alias{areEventConfigurationsIdentical}
\alias{areShiftSetsEqual}
\alias{as.bammdata}
\alias{as.bammdata.credibleshiftset}
\alias{as.phylo.bammdata}
\alias{branchMeanRateExponential}
\alias{colorMap}
\alias{exponentialRate}
\alias{getPathToRoot}
\alias{getRecursiveSequence}
\alias{getSampleCoMat}
\alias{getSequenceForwardTraversal}
\alias{getStartStopTimes}
\alias{rateLegend}
\alias{inv.logit}
\alias{logit}
\alias{mkdtsegsPolar}
\alias{mkdtsegsPhylo}
\alias{richColors}
\alias{recursivelySetNodeStates}
\alias{segMap}
\alias{timeIntegratedBranchRate}
\alias{transparentColor}
\alias{getDesc}
\alias{getSpanningTaxonPair}
\alias{print.bammdata}
\alias{setPhyloTreeCoords}
\alias{setPolarTreeCoords}
\alias{phylogeneticMean}

\title{Internal BAMMtools functions}
\description{
\code{BAMMtools-internal} functions are not routinely called by the user and are undocumented.
}

\details{
Internal \pkg{BAMMtools} functions, either not intended to be called directly by the user or otherwise undocumented.
}
 