\name{BAMMtools-package}
\alias{BAMMtools-package}
\alias{BAMMtools}
\docType{package}
\title{
BAMMtools
}
\description{
An R package for the analysis and visualization of complex macroevolutionary dynamics. Functions in \code{BAMMtools} are oriented entirely around analysis of results obtained using the \code{BAMM} software (\code{bamm-project.org}).

}
\details{
\tabular{ll}{
Package: \tab BAMMtools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-05\cr
License: \tab GPL-2\cr
}

}

\author{
Dan Rabosky, Carlos Anderson, Mike Grundler, Pascal Title, Jeff Shi, Joseph Brown

Maintainer: Dan Rabosky <dan.rabosky@gmail.com>
}
\references{
\url{bamm-project.org}
}

\keyword{ package }
