\name{BAMMtools-package}
\alias{BAMMtools-package}
\alias{BAMMtools}
\docType{package}
\title{
BAMMtools
}
\description{
An R package for the analysis and visualization of complex macroevolutionary dynamics. Functions in \code{BAMMtools} are oriented entirely around analysis of results obtained using the \code{BAMM} software (\code{bamm-project.org}).

}
\details{
\tabular{ll}{
Package: \tab BAMMtools\cr
Type: \tab Package\cr
Version: \tab 2.0.6\cr
Date: \tab 2015-09-10\cr
License: \tab GPL-2 | GPL-3\cr
}

}

\author{
Dan Rabosky, Mike Grundler, Pascal Title, Carlos Anderson, Jeff Shi, Joseph Brown, Huateng Huang\cr
Maintainer: {Pascal Title <ptitle@umich.edu}>
}
\references{
\url{bamm-project.org}

Rabosky, D., M. Grundler, C. Anderson, P. Title, J. Shi, J. Brown, H. Huang and J. Larson. 2014. BAMMtools: an R package for the analysis of evolutionary dynamics on phylogenetic trees. Methods in Ecology and Evolution 5: 701-707.

Rabosky, D. L. 2014. Automatic detection of key innovations, rate shifts, and diversity-dependence on phylogenetic trees. PLoS ONE 9: e89543

Shi, J. J., and D. L. Rabosky. 2015. Speciation dynamics during the global radiation of extant bats. Evolution 69: 1528-1545.

Rabosky, D. L., F. Santini, J. T. Eastman, S. A. Smith, B. L. Sidlauskas, J. Chang, and M. E. Alfaro. 2013. Rates of speciation and morphological evolution are correlated across the largest vertebrate radiation. Nature Communications DOI: 10.1038/ncomms2958 

}

\keyword{ package }
