\name{summary.branchprior}
\alias{summary.branchprior}
\alias{print.branchprior}
\title{
Summary of class "branchprior" object.
}
\description{
Prints summary attributes of the \code{branchprior} object 
}
\usage{
\method{summary}{branchprior}(object, ...)
\method{print}{branchprior}(x, ...)
}
\arguments{
  \item{object,x}{
an object of class \code{branchprior}.
}
\item{...}{
	additional arguments (unused).
}
}


\references{
\url{bamm-project.org}
}
\author{
Dan Rabosky
}

\seealso{
\code{\link{distinctShiftConfigurations}}, \code{\link{plot.bammshifts}}, \code{\link{credibleShiftSet}}\code{\link{getBranchShiftPriors}}
} 
 
 
\keyword{ models }
